update to multifish driver...
Please, no mention in the whatsnew

adds and documents all known sets (3 years or old).

Warning: Some sets have the version hacked to show sets "newer" then 3 years old.  But these have been verfied to be older sets.  There is also a note in the driver about version changes being made to make the opperator look like they are running a version that meets the standard of law change in 2007