Mario - added proper mixer to final output.  This adds the missing filtering on the walking sounds.
This following is not worth mentioning yet.
Started the new 74LS624 code.  Called it DISCRETE_74LS629 for now, until it is ready to replace the old code.  You can enable it in dkongjr and mario using the USE_LS629 define.

I still need to work out the threshold/range relationship.  I'm having the old schrodinger's cat problem.  I need to measure the cap threshold, but doing so changes it.  I'll work it out.