taito_1.c - Added XTALs to driver, fixing MT bug #03598.
Comment:  Added a MACHINE_START for LAGIRL, which doesn't want to work properly (assumes I have service mode ON?)  The working method is still in place (using cachat's startup).  The clock changes were per the source notes above the ROM_START.  If anyone wants to get it working with the proper speeds, be my guest.