First round of an attempted cleanup of header files in the system.  - Created new central header "emu.h"; this should be included     by pretty much any driver or device as the first include. This     file in turn includes pretty much everything a driver or device     will need, minus any other devices it references. Note that     emu.h should *never* be included by another header file.  - Updated all files in the core (src/emu) to use emu.h.  - Removed a ton of redundant and poorly-tracked header includes     from within other header files.  - Temporarily changed driver.h to map to emu.h until we update     files outside of the core.
Added class wrapper around tagmap so it can be directly included
and accessed within objects that need it. Updated all users to
embed tagmap objects and changed them to call through the class.

Added nicer functions for finding devices, ports, and regions in
a machine:

   machine->device("tag") -- return the named device, or NULL
   machine->port("tag") -- return the named port, or NULL
   machine->region("tag"[, &length[, &flags]]) -- return the
      named region and optionally its length and flags
      
Made the device tag an astring. This required touching a lot of 
code that printed the device to explicitly fetch the C-string
from it. (Thank you gcc for flagging that issue!)