Adds projection calculation for 3d eye-space points. [Andrew Gardner]
* out of whatsnew.txt *
This code is to just show a decent method for projecting the 3d points into 2d.
The code itself is fine, but the way it splats points to the screen is silly and
Angelo will likely know a much better way to hook it up.
It looks like the focal length of the camera is a little short right now, but
we'll find the right values eventually.