Rewrote SAMPLES as a modern device. Updated all callers. FLAC reading is now done using the FLAC wrapper. There is now a samples_iterator class to centralize the logic for handling the sample list walking.
Also redid the cheesy half-baked votrax device since it relied 
on some old samples-based handling. Until we have a real 
implementation, it would be good to route the various clients 
through the current one to at least wire it up properly, even 
if it just plays samples in the end. Will look into that 
shortly.