Pong now uses netlist.h. This implementation reproduces all gate timing and 555 related effects as described by Dr. H. Holden in his 2012 publication "ATARI PONG E CIRCUIT ANALYSIS & LAWN TENNIS". [Couriersud]
There are now two game entries:
a) pongf
This illustrates how a workaround has to be used in the absence of gate timing. The game uses a code based netlist.
b) pong
This uses gate timing at a resolution of approx. 140 MHz. The netlist itself is read from the rom region.