a regular 68000 processor will generate a read operation when using the clr.b/.w/.l opcodes, this was fixed / changed on the 020.
I'm not aware of any software explicitly requiring this right now, but it would be detectable if you have read-triggered hardware and various 68k documents advise caution when using the opcodes / porting code using said opcodes to a 020 based machine for that reason.

There was some private discussion with Charles before doing this, apparently it is detectable on the Genesis where a read from the VDP port will increment the internal address counter.