added generic cartslot / ROM socket slot device, which offers basic allocation and access handlers, and converted a few drivers to use this instead of code from cartslot.c [Fabio Priuli]
out of whatsnew: the RAM socket part is just a proof of concept,
and the natural extension of the line of thought which lead me
to this generic socket/cartslot. it might allow to convert current RAM 
device to be a slot device as well (after some refactorization, of 
course, since current code lacks many of the necessary features), 
or be removed soonish, depending on consensus.